Update to 2.45, it seems make check runs?
If you are reading this and the check step is failing for you, please
open a comment in AUR. Try to put || true in the check() step again and
see if it works. Hey, at least in my computer works although I should
test in a clean Arch Linux environment.